Rename `PexInterpreterConstraints` to `InterpreterConstraints` and move to dedicated file
