+++ src/python/pants/backend/awslambda/python/rules.py
@@ -12,8 +12,8 @@ from pants.backend.awslambda.python.target_types import (
     ResolvePythonAwsHandlerRequest,
 )
 from pants.backend.python.util_rules import pex_from_targets
+from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
 from pants.backend.python.util_rules.pex import (
-    PexInterpreterConstraints,
     PexPlatforms,
     PexRequest,
     PexRequirements,
@@ -96,7 +96,7 @@ async def package_python_awslambda(
         output_filename="lambdex.pex",
         internal_only=True,
         requirements=PexRequirements(lambdex.all_requirements),
-        interpreter_constraints=PexInterpreterConstraints(lambdex.interpreter_constraints),
+        interpreter_constraints=InterpreterConstraints(lambdex.interpreter_constraints),
         main=lambdex.main,
     )
