+++ src/python/pants/backend/python/lint/bandit/rules.py
@@ -8,13 +8,8 @@ from pants.backend.python.lint.bandit.skip_field import SkipBanditField
 from pants.backend.python.lint.bandit.subsystem import Bandit
 from pants.backend.python.target_types import InterpreterConstraintsField, PythonSources
 from pants.backend.python.util_rules import pex
-from pants.backend.python.util_rules.pex import (
-    PexInterpreterConstraints,
-    PexRequest,
-    PexRequirements,
-    VenvPex,
-    VenvPexProcess,
-)
+from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
+from pants.backend.python.util_rules.pex import PexRequest, PexRequirements, VenvPex, VenvPexProcess
 from pants.core.goals.lint import LintReport, LintRequest, LintResult, LintResults, LintSubsystem
 from pants.core.util_rules.config_files import ConfigFiles, ConfigFilesRequest
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
@@ -47,7 +42,7 @@ class BanditRequest(LintRequest):
 @dataclass(frozen=True)
 class BanditPartition:
     field_sets: Tuple[BanditFieldSet, ...]
-    interpreter_constraints: PexInterpreterConstraints
+    interpreter_constraints: InterpreterConstraints
 
 
 def generate_argv(
@@ -138,7 +133,7 @@ async def bandit_lint(
     # ( https://github.com/PyCQA/bandit#under-which-version-of-python-should-i-install-bandit). We
     # batch targets by their constraints to ensure, for example, that all Python 2 targets run
     # together and all Python 3 targets run together.
-    constraints_to_field_sets = PexInterpreterConstraints.group_field_sets_by_constraints(
+    constraints_to_field_sets = InterpreterConstraints.group_field_sets_by_constraints(
         request.field_sets, python_setup
     )
     partitioned_results = await MultiGet(
